Feature add code

+++ LuckyDrawPromotion/Services/AwardService.cs
@@ -1,4 +1,5 @@
 ﻿using LuckyDrawPromotion.Data.Entity;
+using LuckyDrawPromotion.Models;
 using LuckyDrawPromotion.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -19,19 +20,52 @@ namespace LuckyDrawPromotion.Services
             var awardIsExist = await _context.Awards.FirstOrDefaultAsync(x => x.Id == id);
             if (awardIsExist == null)
                 return false;
+            var code = await _context.Codes.FirstOrDefaultAsync(x => x.Id == awardIsExist.IdCode);
+            if (code.UsedCount == 0)
+                return false;
+            code.UsedCount--;
             _context.Awards.Remove(awardIsExist);
+            _context.Codes.Update(code);
             await _context.SaveChangesAsync();
             return true;
         }
 
-        public async Task<IEnumerable<Award>?> Get()
+        public async Task<List<ListWinnerViewModel>> GetListAwardAsync(string phoneNumber, string nameCampaign)
         {
-            return await _context.Awards.ToListAsync();
-        }
-
-        public async Task<Award?> Get(Guid id)
-        {
-            return await _context.Awards.FirstOrDefaultAsync(x => x.Id == id);
+            var campaign = await _context.Campaigns.FirstOrDefaultAsync(x => x.Name == nameCampaign);
+            if (campaign == null)
+                return new List<ListWinnerViewModel>();
+            if (string.IsNullOrEmpty(phoneNumber))
+            {
+                var listCode = await (from c in _context.Codes
+                                      join g in _context.Gifts on c.IdGift equals g.Id
+                                      where c.NameCampaign == nameCampaign && g.Name != "Chúc bạn may mắn lần sau"
+                                      select c).ToListAsync();
+                var listWinner = (from c in listCode
+                                  join a in _context.Awards on c.Id equals a.IdCode
+                                  orderby a.UsedDate descending
+                                  select new ListWinnerViewModel
+                                  {
+                                      PhoneNumber = a.PhoneNumberUser,
+                                      Date = Convert.ToDateTime(a.UsedDate).ToLocalTime(),
+                                      Award = _context.Gifts.First(x => x.Id == c.IdGift).Name
+                                  }).ToList();
+                return listWinner;
+            }
+            var user = await _context.Users.FirstOrDefaultAsync(x => x.PhoneNumber == phoneNumber);
+            if (user == null)
+                return new List<ListWinnerViewModel>();
+            var listWinners = (from c in _context.Codes
+                              join a in _context.Awards on c.Id equals a.IdCode
+                              where a.PhoneNumberUser == phoneNumber && c.NameCampaign == nameCampaign
+                              orderby a.UsedDate descending
+                              select new ListWinnerViewModel
+                              {
+                                  PhoneNumber = a.PhoneNumberUser,
+                                  Date = Convert.ToDateTime(a.UsedDate).ToLocalTime(),
+                                  Award = _context.Gifts.First(x => x.Id == c.IdGift).Name
+                              }).ToList();
+            return listWinners;
         }
 
         public async Task<bool> Post([FromBody] Award award)
@@ -42,6 +76,10 @@ namespace LuckyDrawPromotion.Services
             if (userIsExist == null || codeIsExist == null || awardIsExist != null)
                 return false;
 
+            if (codeIsExist.UsedCount == codeIsExist.LimitUsage)
+                return false;
+            codeIsExist.UsedCount++;
+
             if (award.UsedDate != null)
                 if (award.UsedDate > DateTime.Now)
                     return false;
@@ -51,32 +89,7 @@ namespace LuckyDrawPromotion.Services
                 }
 
             await _context.Awards.AddAsync(award);
-            await _context.SaveChangesAsync();
-            return true;
-        }
-
-        public async Task<bool> Put(Guid id, [FromBody] Award award)
-        {
-            var awardIsExist = await _context.Awards.FindAsync(award.Id);
-            if (awardIsExist == null)
-                return false;
-
-            _context.Remove(awardIsExist);
-            await _context.SaveChangesAsync();
-
-
-            awardIsExist = await _context.Awards.FindAsync(award.Id);
-            var checkIsExist = await _context.Awards.FirstOrDefaultAsync(x => x.IdCode == award.IdCode
-                                                                           && x.PhoneNumberUser == award.PhoneNumberUser
-                                                                           && x.UsedDate == award.UsedDate);
-            if (awardIsExist != null || checkIsExist != null)
-                return false;
-
-            if (award.UsedDate > DateTime.Now)
-                return false;
-            award.UsedDate = DateTime.UtcNow;
-
-            await _context.Awards.AddAsync(award);
+            _context.Codes.Update(codeIsExist);
             await _context.SaveChangesAsync();
             return true;
         }
